lead-gen-agent/requirements.txt
google-generativeai
duckduckgo-search
beautifulsoup4
requests
pandas
python-dotenv
